# doc/src/main/java/com/example/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnlineShop API",
    "description": "java onlineshop",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://127.0.0.1",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/panel/user/{id}": {
      "get": {
        "tags": [
          "user-panel-controller"
        ],
        "operationId": "getById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseUserDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user-panel-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/invoice/{id}": {
      "get": {
        "tags": [
          "invoice-panel-controller"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseInvoiceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/invoice/user/{uid}": {
      "get": {
        "tags": [
          "invoice-panel-controller"
        ],
        "operationId": "getAllById",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseListInvoiceDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/invoice/mine/{id}": {
      "get": {
        "tags": [
          "invoice-panel-controller"
        ],
        "operationId": "getMineInfo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseInvoiceDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponseUserDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "CustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "tel": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          }
        }
      },
      "PermissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "RoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionDto"
            },
            "uniqueItems": true
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "registerDate": {
            "type": "string",
            "format": "date-time"
          },
          "enable": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            },
            "uniqueItems": true
          },
          "customerDto": {
            "$ref": "#/components/schemas/CustomerDto"
          }
        }
      },
...
      "APIResponseBoolean": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "boolean"
          }
        }
      },
      "File": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SizeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "ColorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "hex": {
            "type": "string"
          }
        }
      },
      "FileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ProductCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "enable": {
            "type": "boolean"
          }
        }
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enable": {
            "type": "boolean"
          },
          "exist": {
            "type": "boolean"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "addDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "colors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorDto"
            },
            "uniqueItems": true
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SizeDto"
            },
            "uniqueItems": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategoryDto"
          }
        }
      },
...
      "APIResponseInvoiceDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/InvoiceDto"
          }
        }
      },
      "Color": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "hex": {
            "type": "string"
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "tel": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "payedDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "Payed",
              "InProgress"
            ]
          },
          "totalAmount": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItem"
            }
          }
        }
      },
      "InvoiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "payedDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "Payed",
              "InProgress"
            ]
          },
          "totalAmount": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItemDto"
            }
          }
        }
      },
      "InvoiceItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "invoice": {
            "$ref": "#/components/schemas/Invoice"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "size": {
            "$ref": "#/components/schemas/Size"
          },
          "color": {
            "$ref": "#/components/schemas/Color"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "InvoiceItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "invoice": {
            "$ref": "#/components/schemas/Invoice"
          },
          "product": {
            "$ref": "#/components/schemas/ProductDto"
          },
          "size": {
            "$ref": "#/components/schemas/SizeDto"
          },
          "color": {
            "$ref": "#/components/schemas/ColorDto"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "parent": {
            "$ref": "#/components/schemas/Permission"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enable": {
            "type": "boolean"
          },
          "exist": {
            "type": "boolean"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "addDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "$ref": "#/components/schemas/File"
          },
          "colors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Color"
            },
            "uniqueItems": true
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Size"
            },
            "uniqueItems": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategory"
          }
        }
      },
      "ProductCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enable": {
            "type": "boolean"
          },
          "image": {
            "$ref": "#/components/schemas/File"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "uniqueItems": true
          }
        }
      },
      "Size": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "registerDate": {
            "type": "string",
            "format": "date-time"
          },
          "enable": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "uniqueItems": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          }
        }
      },
      "APIResponseListInvoiceDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDto"
            }
          }
        }
      },
...
    }
  }
}